create did_model submodule
