+++ doubleml/__init__.py
@@ -5,8 +5,8 @@ from .iv_model.pliv import DoubleMLPLIV
 from .double_ml_irm import DoubleMLIRM
 from .iv_model.iivm import DoubleMLIIVM
 from .double_ml_data import DoubleMLData, DoubleMLClusterData
-from .double_ml_did import DoubleMLDID
-from .double_ml_did_cs import DoubleMLDIDCS
+from .did_model.did import DoubleMLDID
+from .did_model.did_cs import DoubleMLDIDCS
 from .double_ml_qte import DoubleMLQTE
 from .double_ml_pq import DoubleMLPQ
 from .double_ml_lpq import DoubleMLLPQ
