+++ doubleml/did_model/did_cs.py
@@ -3,12 +3,12 @@ from sklearn.utils import check_X_y
 from sklearn.utils.multiclass import type_of_target
 import warnings
 
-from .double_ml import DoubleML
-from .double_ml_data import DoubleMLData
-from .double_ml_score_mixins import LinearScoreMixin
+from ..double_ml import DoubleML
+from ..double_ml_data import DoubleMLData
+from ..double_ml_score_mixins import LinearScoreMixin
 
-from .utils._estimation import _dml_cv_predict, _trimm, _get_cond_smpls_2d, _dml_tune
-from .utils._checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
+from ..utils._estimation import _dml_cv_predict, _trimm, _get_cond_smpls_2d, _dml_tune
+from ..utils._checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
 
 
 class DoubleMLDIDCS(LinearScoreMixin, DoubleML):
