+++ doubleml/did_model/tests/test_did.py
@@ -9,7 +9,7 @@ from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
 
 import doubleml as dml
 
-from ._utils import draw_smpls
+from ...tests._utils import draw_smpls
 from ._utils_did_manual import fit_did, boot_did, fit_sensitivity_elements_did
