+++ doubleml/did_model/tests/test_did_tune.py
@@ -9,7 +9,7 @@ from sklearn.ensemble import RandomForestRegressor
 
 import doubleml as dml
 
-from ._utils import draw_smpls
+from ...tests._utils import draw_smpls
 from ._utils_did_manual import fit_did, boot_did, tune_nuisance_did
